requirements.txt
numpy==1.26.4
opencv-python==4.10.0.82
pika==1.3.2
pillow==10.3.0
